Pomodoro: extended pause sessions and sound
After four consecutive work sessions, the counter sets an extended pause
session of the same length of a work session.
Sound is played at the end of both work and pause sessions.
Using a promise to deal with a case where the sound doesn't play. This
happens for example in Safari 11, where auto playback is blocked.